requirements.txt
amqp==5.0.1
asgiref==3.2.10
billiard==3.6.3.0
celery==5.0.0
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1.1
django-celery-beat==2.0.0
django-filter==2.4.0
django-timezone-field==4.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
kombu==5.0.2
prompt-toolkit==3.0.7
psycopg2==2.8.6
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
vine==5.0.0
wcwidth==0.2.5
